Moude Guide
This mouse guide was invented in a coding session the night before Christmas in December of 2020 when my hands were aching and everyone in the Talon slack was coming up with ideas for a better mouse.  This code was all Ryan's work, (along with the whole talon system itself.)  I have used it ever since.